moving EVERYTHING
