+++ package.json
@@ -7,13 +7,18 @@
     "@types/bun": "latest",
     "eslint": "^9.27.0"
   },
+  "scripts": {
+    "start": "node src/bot/index.js",
+    "deploy-commands": "node deploy-commands.js",
+    "dev": "node src/bot/index.js --watch",
+    "lint": "eslint . --ext .js,.ts"
+  },
   "peerDependencies": {
     "typescript": "^5"
   },
   "dependencies": {
     "@balldontlie/sdk": "^1.2.2",
     "@napi-rs/canvas": "^0.1.70",
-    "bun": "^1.2.13",
     "discord.js": "^14.19.3",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
