+++ src/assets/crown.png
