+++ src/bot/commands/fun/deathbattle.js
