+++ src/bot/commands/fun/deathbattle.json
