+++ src/bot/commands/fun/fnshop.js
