+++ src/bot/commands/fun/fnstats.js
