+++ src/bot/commands/fun/lolprofile.js
