+++ src/bot/commands/fun/madlibs.js
