+++ src/bot/commands/fun/madlibs.json
