+++ src/bot/commands/fun/searchbook.js
