+++ src/bot/commands/fun/tony.js
