+++ src/bot/commands/fun/votetimeout.js
@@ -6,7 +6,7 @@ const {
   EmbedBuilder,
   MessageFlags,
 } = require("discord.js");
-const { Configurations } = require("../../models.js");
+const { Configurations } = require("../../../models.js");
 
 module.exports = {
   data: new SlashCommandBuilder()
