+++ src/bot/commands/fun/waifu.js
