+++ src/bot/commands/utility/daily.js
