+++ src/bot/commands/utility/ping.js
