+++ src/bot/commands/utility/reactionRoleAdd.js
@@ -1,25 +1,29 @@
 const { MessageFlags, SlashCommandBuilder } = require('discord.js');
-const { ReactionRoles } = require('../../models.js');
+const { ReactionRoles } = require('../../../models.js');
 
 module.exports = {
     data: new SlashCommandBuilder()
-        .setName('reactionroleremove')
-        .setDescription('Remove a reaction role from a message.')
+        .setName('reactionroleadd')
+        .setDescription('Add a reaction role to a message.')
+
         .addStringOption(option =>
             option.setName('messageid')
-                .setDescription('The message id for the role to be removed.')
+                .setDescription('the message id for the role to be added.')
                 .setRequired(true),
         )
         .addStringOption(option =>
             option.setName('emoji')
                 .setDescription('The emoji for the role')
                 .setRequired(true)
+
         )
         .addRoleOption(option =>
             option.setName('role')
-                .setDescription('The role of which you are removing.')
+                .setDescription('the role of which you are reacting to.')
                 .setRequired(true),
         ),
+
+
 
     async execute(interaction) {
         const message_id = interaction.options.getString('messageid');
@@ -29,7 +33,7 @@ module.exports = {
         try {
             const message = await interaction.channel.messages.fetch(message_id);
 
-            const reaction_role = await ReactionRoles.findOne({
+            let reaction_role = await ReactionRoles.findOne({
                 where: {
                     messageid: message_id,
                     guildid: interaction.guild.id,
@@ -38,39 +42,41 @@ module.exports = {
                 },
             });
 
-            if (reaction_role) {
-                await reaction_role.destroy();
+            if (!reaction_role) {
+                reaction_role = await ReactionRoles.create({ 
+                    messageid: message_id,
+                    guildid: interaction.guild.id,
+                    roleid: role_id,
+                    emoji: emoji,
+                });
 
-                const reaction = message.reactions.cache.get(emoji) ||
-                    message.reactions.cache.find(r => r.emoji.toString() === emoji);
-                if (reaction) {
-                    await reaction.users.remove(interaction.client.user);
-                }
-
+                await message.react(emoji);
                 await interaction.reply({
-                    content: 'reaction role removed',
+                    content: 'reaction role added',
                     flags: MessageFlags.Ephemeral,
                 });
             }
             else {
                 await interaction.reply({
-                    content: 'no reaction role found',
+                    content: 'reaction role already added',
                     flags: MessageFlags.Ephemeral,
                 });
             }
-
         }
         catch (error) {
-            if (error.code === 10014) {
+            if (error === 10014) {
                 await interaction.reply({
                     content: 'emoji not found',
                     flags: MessageFlags.Ephemeral,
                 });
             }
-            await interaction.reply({
-                content: 'issue removing reaction role',
-                flags: MessageFlags.Ephemeral,
-            });
+            else {
+                await interaction.reply({
+                    content: 'issues adding reactions',
+                    flags: MessageFlags.Ephemeral,
+                });
+            }
+
         }
-    },
-};
+    }
+}
