+++ src/bot/commands/utility/removeWelcomeLog.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder, MessageFlags } = require('discord.js');
-const { Configurations } = require('../../models.js');
+const { Configurations } = require('../../../models.js');
 
 module.exports = {
     data: new SlashCommandBuilder()
