+++ src/bot/commands/utility/user.js
