+++ src/bot/events/guildMemberRemove.js
@@ -1,5 +1,5 @@
 const { Events, EmbedBuilder } = require('discord.js');
-const { Configurations } = require('../models.js');
+const { Configurations } = require('../../models.js');
 
 module.exports = {
     name: Events.GuildMemberRemove,
