+++ src/bot/events/messageDelete.js
